feature(contacts-nodemailer): fixing mailController
