+++ server/controllers/mailController.js
@@ -1,25 +1,40 @@
+const nodemailer = require('nodemailer');
 const dotenv = require('dotenv');
-const mailServer = require('../services/sendMessage');
+
+dotenv.config();
 
 const sendMail = async (req, res) => {
   try {
     const { name, email, phone, topic, message } = req.body;
+
+    const transporter = nodemailer.createTransport({
+      host: 'smtp.gmail.com',
+      port: 587,
+      secure: false,
+      auth: {
+        user: process.env.EMAIL,
+        pass: process.env.PASSWORD,
+      },
+      tls : { rejectUnauthorized: false }
+    });
+
+    /*console.log("1 2 3");*/
+    await transporter.sendMail({
+      from: process.env.EMAIL,
+      to: process.env.EMAILTO,
+      subject: topic,
+      html: `<p> Nome: ${name} </p>
+      <p> Telefone: ${phone} </p>
+      <p> Email: ${email} </p>
+      <p> Assunto: ${topic} </p>
+      <p> Mensagem: ${message} </p>`,
+    });
 
     const body = `Nome: ${name}
     Email: ${email}
     Telefone: ${phone}
     Assunto: ${topic}
     Mensagem: ${message}`;
-
-    await mailServer({
-      destinationUser: process.env.EMAILTO,
-      subjectText: topic,
-      htmlOption:`<p> Nome: ${name} </p>
-      <p> Telefone: ${phone} </p>
-      <p> Email: ${email} </p>
-      <p> Assunto: ${topic} </p>
-      <p> Mensagem: ${message} </p>`,
-    });
 
     res.status(200).send("Email sent")
