--- server/services/sendMessage.js
@@ -1,29 +1,0 @@
-const nodemailer = require('nodemailer');
-const dotenv = require('dotenv');
-
-dotenv.config();
-
-module.exports = ( async (
-  destinationUser,
-  subjectText,
-  htmlOption,
-) => {
-  const transporter = nodemailer.createTransport({
-    host: 'smtp.gmail.com',
-    port: 587,
-    secure: false,
-    auth: {
-      user: process.env.EMAIL,
-      pass: process.env.PASSWORD,
-    },
-    tls : { rejectUnauthorized: false }
-  });
-
-  /*console.log("1 2 3");*/
-  await transporter.sendMail({
-    from: process.env.EMAIL,
-    to: process.env.EMAILTO,
-    subject: subjectText,
-    html: htmlOption,
-  });
-});
